backend/requirements.txt
aniso8601==9.0.1
async-timeout==5.0.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pytz==2024.2
redis==5.2.0
six==1.16.0
Werkzeug==3.1.2
zipp==3.20.2
